Unificate the item component
